+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+ - CrestronFusion - RVI generation now runs in it's own low-priority thread
+ 
 ## [5.4.0] - 2020-03-20
 ### Added
  - Added error catching around adding sigs and assets to fusion to better log what sigs are causing errors.